+++ CHANGES.md
@@ -3,6 +3,7 @@
 
 ## 0.12.0
 * major: Completely remove Swagger 1.2 support
+* major: Remove working within the browser
 
 ## 0.11.0
 * major: changed module name to @axway/swagger-tools